turtle 3

+++ src/main/java/canvas/MyCanvas.java
@@ -16,8 +16,8 @@ public class MyCanvas extends JPanel {
     }
 
     public void paintComponent(Graphics g){
-        g.fillOval(turtle.getX()-15, turtle.getY()-15, 30,30);
+        g.fillOval((int) (turtle.getX()-15), (int) (turtle.getY()-15), 30,30);
         g.setColor(Color.red);
-        g.fillOval(turtle.getX()+10, turtle.getY()-5, 10,10);
+        g.fillOval((int) (turtle.getX()+10), (int) (turtle.getY()-5), 10,10);
     }
 }
